tracks/track-06-standard-agent-prompts/requirements.txt
# Standard Agent framework
standard-agent>=0.1.0

# Jentic SDK for API access
jentic>=0.1.0

# Environment management
python-dotenv>=1.0.0

# Testing and validation
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Data handling
pydantic>=2.0.0
jsonschema>=4.0.0

# Optional: Additional utilities
rich>=13.0.0  # For pretty console output
click>=8.0.0  # For CLI tools